A few ubuntu fixes

+++ src/CPU.h
@@ -8,8 +8,8 @@
 #define BYTES_PER_LETTER 5
 #define SCR_W 64
 #define SCR_H 32
-#define FPS 60
-#define TICKS_PER_FRAME (1000/FPS)
+#define FPS ((float)60)
+#define TICKS_PER_FRAME ((uint32_t)(1000/FPS))
 
 typedef struct
 {
